refactor(flat-table): move context into an internal directory
Moves `FlatTableHeadContext` into an `__internal__` directory and updates imports where
neccessary. The context is not intended for public use and should improve
tree-shaking in Carbon. Renames `FlatTableRowContext` file to match naming convention

BREAKING CHANGE: The context is no longer exposed as part of the public
interface of `FlatTableHead`
