+++ src/components/flat-table/flat-table-head/__internal__/flat-table-head.context.ts
@@ -1,0 +1,9 @@
+import React from "react";
+
+interface FlatTableHeadContextProps {
+  stickyOffsets: Record<string, number>;
+}
+
+export default React.createContext<FlatTableHeadContextProps>({
+  stickyOffsets: {},
+});
