+++ src/components/flat-table/flat-table-head/flat-table-head.component.tsx
@@ -1,21 +1,12 @@
 import React, { useEffect, useState, useRef } from "react";
 import StyledFlatTableHead from "./flat-table-head.style";
 import { buildPositionMap } from "../__internal__";
+import FlatTableHeadContext from "./__internal__/flat-table-head.context";
 
 export interface FlatTableHeadProps {
   /** Array of FlatTableRow. */
   children: React.ReactNode;
 }
-
-interface FlatTableHeadContextProps {
-  stickyOffsets: Record<string, number>;
-}
-
-export const FlatTableHeadContext = React.createContext<FlatTableHeadContextProps>(
-  {
-    stickyOffsets: {},
-  }
-);
 
 export const FlatTableHead = ({ children, ...rest }: FlatTableHeadProps) => {
   const ref = useRef<HTMLTableSectionElement>(null);
