+++ src/components/flat-table/flat-table-row-header/flat-table-row-header.spec.tsx
@@ -10,7 +10,7 @@ import {
   testStyledSystemPadding,
 } from "../../../__spec_helper__/__internal__/test-utils";
 import StyledIcon from "../../icon/icon.style";
-import FlatTableRowContext from "../flat-table-row/__internal__/flat-table-row-context";
+import FlatTableRowContext from "../flat-table-row/__internal__/flat-table-row.context";
 
 describe("FlatTableRowHeader", () => {
   testStyledSystemPadding(
