+++ src/components/flat-table/flat-table-row/flat-table-row.component.tsx
@@ -18,10 +18,10 @@ import FlatTableRowDraggable, {
 } from "./__internal__/flat-table-row-draggable.component";
 import FlatTableContext from "../__internal__/flat-table.context";
 import guid from "../../../__internal__/utils/helpers/guid";
-import FlatTableRowContext from "./__internal__/flat-table-row-context";
+import FlatTableRowContext from "./__internal__/flat-table-row.context";
 import SubRowProvider, { SubRowContext } from "./__internal__/sub-row-provider";
 import { buildPositionMap } from "../__internal__";
-import { FlatTableHeadContext } from "../flat-table-head/flat-table-head.component";
+import FlatTableHeadContext from "../flat-table-head/__internal__/flat-table-head.context";
 
 export interface FlatTableRowProps {
   /** Overrides default cell color, provide design token, any color from palette or any valid css color value. */
